+++ source/renderer/util/table-helper.js
@@ -154,7 +154,9 @@ class TableHelper {
    * @return {void}               Does not return.
    */
   fromMarkdown (markdownTable) {
-    // This function re-builds the full table from a given Markdown table.
+    // First remove whitespace from both sides of the table, e.g. in case
+    // a trailing newline is present
+    markdownTable = markdownTable.trim()
     if (!Array.isArray(markdownTable)) markdownTable = markdownTable.split('\n')
 
     if (markdownTable.length === 0) throw new Error('MarkdownTable was empty!')
@@ -163,35 +165,47 @@ class TableHelper {
     let ast = [] // Two-dimensional array
     let colAlignments = [] // One-dimensional column alignments
     let numColumns // If there is an uneven number of columns, throw an error.
+
     // Now iterate over all table rows
     for (let i = 0; i < markdownTable.length; i++) {
-      // Ignore empty lines
-      if (markdownTable[i].trim() === '') continue
-      let row = markdownTable[i]
-
-      // Remove surrounding pipes
-      if (row[0] === '|') row = row.substr(1)
-      if (row[row.length - 1] === '|') row = row.substr(0, row.length - 1)
+      // There should not be empty lines in the table.
+      // If so, this indicates an error in the render tables plugin!
+      if (markdownTable[i].trim() === '') throw new Error(`Line ${i} in the table was empty!`)
+      let row = markdownTable[i].trim() // Clean up whitespace
 
       // Split to columns
       row = row.split('|')
 
+      // Now, expect that the first and last "column" are empty
+      // and remove them.
+      if (row[0].trim() !== '' || row[row.length - 1].trim() !== '') {
+        throw new Error(`Malformed Markdown table! Row ${i} was not surrounded by whitespace!`)
+      }
+
+      row.pop() // Out goes the last ...
+      row.shift() // ... and the first
+
+      console.log('Parsing row ...', row)
+
       // First row determines the amount of columns expected
       if (!numColumns) numColumns = row.length
-      if (numColumns !== row.length) throw new Error('Malformed Markdown Table!')
+      if (numColumns !== row.length) {
+        throw new Error(`Malformed Markdown Table! Found ${row.length} columns rows on line ${i} (should be ${numColumns}).`)
+      }
 
-      // First test if we've got a header row.
-      // A header row is defined of consisting of columns only containing
-      // dashes and colons. The first column to break this rule means we
-      // don't have a valid header.
+      // First test if we've got a header row. A header row is defined of
+      // consisting of columns only containing dashes, colons and spaces. The
+      // first column to break this rule means we don't have a valid header.
       let isHeader = true
       for (let col of row) {
-        if (!/^[ -:]+$/.test(col)) {
+        if (!/^[ -:]+$/.test(col) || col.trim() === '') {
+          // Note we have to check for completely blank lines
           isHeader = false
           break
         }
       }
 
+      // Parse the header
       if (isHeader) {
         for (let j = 0; j < row.length; j++) {
           let col = row[j].trim()
@@ -206,8 +220,9 @@ class TableHelper {
         continue // We're done here -- don't add it to the AST
       }
 
-      // Now parse all columns
+      // We have a normal table row, so parse all columns
       let cols = []
+      console.log('Parsing ' + row.length + ' cols')
       for (let j = 0; j < row.length; j++) {
         cols.push(row[j].trim()) // Trim whitespaces
       }
@@ -222,6 +237,11 @@ class TableHelper {
       for (let i = 0; i < numColumns; i++) {
         colAlignments.push('left')
       }
+    }
+
+    if (ast.length === 0 || ast[0].length === 0) {
+      // This AST does not look as it's supposed to -> abort
+      throw new Error('Malformed Markdown Table! The parsed AST was empty.')
     }
 
     // Now we need to rebuild everything from the AST